model_compression/requirements.txt
adjustText==1.3.0
matplotlib==3.10.3
model_compression_toolkit==2.2.0
numpy==1.26.4
onnx==1.16.1
Pillow==11.2.1
pycocotools==2.0.8
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
hydra-core==1.3.2
iopath==0.1.10
wandb
